+++ defmt/tests/ui/derive-invalid-attr-arg-item.rs
@@ -1,0 +1,7 @@
+#[derive(defmt::Format)]
+#[defmt(FooBar)]
+struct S {
+    f: bool,
+}
+
+fn main() {}